+++ internal/scorer/fraude_score.go
@@ -1,11 +1,11 @@
 package scorer
 
-func FraudScore(neighbors []SearchResult) float64 {
-	var fraudCount float64
+func FraudScore(neighbors []SearchResult) float32 {
+	var fraudCount float32
 	for _, n := range neighbors {
 		if n.Label == "fraud" {
 			fraudCount++
 		}
 	}
-	return fraudCount / float64(len(neighbors))
+	return fraudCount / float32(len(neighbors))
 }
